refactor: fix "TestStatus" enum
