+++ src/compose/index.ts
@@ -1,2 +1,1 @@
 export { mapScenarios } from "./map-scenarios";
-export { type TestResult } from "./reporting/types";
