--- src/compose/reporting/types.ts
@@ -1,5 +1,0 @@
-export type TestResult = {
-    title: string;
-    status: "succeed" | "failed" | "error" | "not runed" | "canceled";
-    children?: TestResult[];
-};
