+++ src/index.ts
@@ -1,5 +1,6 @@
-export { mapScenarios, type TestResult } from "@compose";
+export { mapScenarios } from "@compose";
 export { createScenario as scenario } from "@modules/scenarios";
 export { createReusableTest as reusableTest } from "@modules/reusable-tests";
 export { listeners } from "@modules/listeners";
 export { expect } from "@core/expectations";
+export { type TestResult } from "@shared/lib";
