+++ src/modules/scenarios/scenario.ts
@@ -1,5 +1,5 @@
 import { TestInfo } from "@modules/tests";
-import { TestStatus } from "@shared/types";
+import { TestStatus } from "@shared/lib";
 
 export class Scenario implements TestInfo {
     public status = TestStatus.NOT_RUNED;
