+++ src/modules/tests/classes/test-group.ts
@@ -1,4 +1,4 @@
-import { TestStatus } from "@shared/types";
+import { TestStatus } from "@shared/lib";
 import { TestInfo } from "../types";
 
 export class TestsGroup implements TestInfo {
