+++ src/modules/tests/types.ts
@@ -1,6 +1,6 @@
 import { TestAborter } from "@core/aborting";
 import { BaseTestFactory } from "./factories/base-test-factory";
-import { TestStatus } from "@shared/types";
+import { TestStatus } from "@shared/lib";
 
 export type TestInfo = {
     title: string;
