+++ src/shared/lib/index.ts
@@ -1,2 +1,3 @@
 export { deepCompare } from "./objects";
 export { Logger } from "./logger";
+export { TestStatus, TestStatusesMap, type TestResult } from "./tests";
