+++ src/shared/types/index.ts
@@ -1,4 +1,3 @@
 export { type TypeofResponse } from "./core";
 export { type ClassConstructor } from "./objects";
 export { type ElementType } from "./arrays";
-export * from "./tests";
