--- src/shared/types/tests.ts
@@ -1,6 +1,0 @@
-export enum TestStatus {
-    SUCCEED = 0,
-    FAILED = 1,
-    NOT_RUNED = 2,
-    ERROR = 3
-}
